requirements.txt
Flask==2.0.1
gunicorn==19.9.0
itsdangerous==2.0
Jinja2>=3.0
MarkupSafe>=2.0
#Werkzeug==0.15.5
scikit-learn==0.24.2
joblib==0.16.0
numpy==1.19.2
Pillow==8.4.0
tensorflow==2.6.0
scipy==1.8.0